project/requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
csscompressor==0.9.5
Django==2.1.7
django-cors-headers==2.5.0
djangorestframework==3.9.2
htmlmin==0.1.12
idna==2.8
image==1.5.27
jsmin==2.2.2
Pillow==5.4.1
pytz==2018.9
requests==2.22.0
soupsieve==1.9.1
urllib3==1.25.3
